+++ lib/coney/consumer/consumer_connection.ex
@@ -1,8 +1,10 @@
 defmodule Coney.ConsumerConnection do
+  alias Coney.ConsumerConnection
+
   defstruct subscribe_channel: nil,
             publish_channel: nil
 
   def build(subscribe_channel, publish_channel) do
-    %Coney.ConsumerConnection{subscribe_channel: subscribe_channel, publish_channel: publish_channel}
+    %ConsumerConnection{subscribe_channel: subscribe_channel, publish_channel: publish_channel}
   end
 end