+++ test/lib/coney/consumer/consumer_executor_test.exs
@@ -1,42 +1,26 @@
 defmodule ConsumerExecutorTest do
   use ExUnit.Case, async: true
 
-  alias Coney.Test.FakeConsumer
   alias Coney.{ConsumerExecutor, ConsumerConnection, ExecutionTask}
 
-  setup do
-    [connection: ConsumerConnection.build(:chan, :chan)]
+  defp build_task(message) do
+    connection = ConsumerConnection.build(:chan, :chan)
+    ExecutionTask.build(FakeConsumer, connection, message, :tag, %{})
   end
 
-  test "consumer failed with error", %{connection: connection} do
-    task = ExecutionTask.build(FakeConsumer, connection, :error, :tag, false)
-
-    result = ConsumerExecutor.consume(task)
-
-    assert :rejected == result
+  test "reject message" do
+    assert :rejected == ConsumerExecutor.consume(build_task(:reject))
   end
 
-  test "consumer failed with exception", %{connection: connection} do
-    task = ExecutionTask.build(FakeConsumer, connection, :exception, :tag, false)
-
-    result = ConsumerExecutor.consume(task)
-
-    assert :failed == result
+  test "consumer failed with exception" do
+    assert :confirmed == ConsumerExecutor.consume(build_task(:exception))
   end
 
-  test "consumer successfully done", %{connection: connection} do
-    task = ExecutionTask.build(FakeConsumer, connection, :ok, :tag, false)
-
-    result = ConsumerExecutor.consume(task)
-
-    assert :confirmed == result
+  test "consumer successfully done" do
+    assert :confirmed == ConsumerExecutor.consume(build_task(:ok))
   end
 
-  test "consumer successfully done with reply", %{connection: connection} do
-    task = ExecutionTask.build(FakeConsumer, connection, :reply, :tag, false)
-
-    result = ConsumerExecutor.consume(task)
-
-    assert :replied == result
+  test "consumer successfully done with reply" do
+    assert :published == ConsumerExecutor.consume(build_task(:reply))
   end
 end